feat(frontend): make Home the surface for empty projects too
A project with no agents got a separate first-run hero: its own page,
its own composer, its own copy, teaching a layout you then never saw
again. Home does that job with one fewer page to learn. It opens on the
templates tab with the composer already describing an agent, and "Your
agents" is not offered until there is one.

So FirstRunScreen, FirstRunComposer, their copy and their states are
deleted, and `homeSurface` moves to the home feature it now serves: it
only decides whether we know enough to draw Home yet. Its two real rules
survive — an error is never read as emptiness, and an empty list that is
still resolving holds rather than flashing the wrong tab.

The tab and the composer's mode are read once at mount, which is when the
answer is known: Home does not render until the list has resolved.
