+++ web/mobile/tests/unit/homeSurface.test.ts
@@ -1,6 +1,6 @@
 import {describe, expect, it} from "vitest"
 
-import {resolveHomeSurface, type HomeSurfaceInput} from "../../src/features/onboarding/homeSurface"
+import {resolveHomeSurface, type HomeSurfaceInput} from "../../src/features/home/homeSurface"
 
 const input = (overrides: Partial<HomeSurfaceInput> = {}): HomeSurfaceInput => ({
     agentCount: 0,
@@ -10,15 +10,15 @@ const input = (overrides: Partial<HomeSurfaceInput> = {}): HomeSurfaceInput => (
 })
 
 describe("resolveHomeSurface", () => {
-    it("gives a settled empty project the first-run hero", () => {
-        expect(resolveHomeSurface(input())).toBe("first-run")
+    it("gives a settled empty project Home, which opens on templates", () => {
+        expect(resolveHomeSurface(input())).toBe("home")
     })
 
-    it("gives a project with agents the normal Home", () => {
+    it("gives a project with agents the same Home", () => {
         expect(resolveHomeSurface(input({agentCount: 3}))).toBe("home")
     })
 
-    it("holds while an empty list is still resolving, so Home never flashes first", () => {
+    it("holds while an empty list is still resolving, so the tabs never flip on arrival", () => {
         expect(resolveHomeSurface(input({isPending: true}))).toBe("loading")
     })
 
@@ -27,14 +27,13 @@ describe("resolveHomeSurface", () => {
     })
 
     it("treats a failed fetch as Home, never as evidence of emptiness", () => {
-        // Desktop's rule: a failed fetch must not send someone who has agents into onboarding,
-        // and Home is the surface that can be retried.
+        // A failed fetch must not be read as "no agents"; Home is the surface that can be retried.
         expect(resolveHomeSurface(input({isError: true}))).toBe("home")
         expect(resolveHomeSurface(input({isError: true, isPending: true}))).toBe("home")
     })
 
-    it("walks a cold first run: pending, then empty", () => {
+    it("walks a cold start: hold while pending, then Home once it settles empty", () => {
         expect(resolveHomeSurface(input({isPending: true}))).toBe("loading")
-        expect(resolveHomeSurface(input())).toBe("first-run")
+        expect(resolveHomeSurface(input())).toBe("home")
     })
 })
